Switch CI back to tox from tox-travis
This reverts a change in 9b572d6e708f8647fae8f31156b19612dc3f3e29.

tox-travis (as previously assumed) doesn't run all the tox envs unless
the env has a python version marker (or a [travis] section is added)
so the previous version would not run the linting. In addition in
travis the tox envs aren't seperated so debuging is harder.